+++ src/main/java/pe/edu/upc/MonolithFoodApplication/repositories/RecipeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.MonolithFoodApplication.entities.RecipeEntity;
 
 @Repository
-public interface RecipeRepository extends JpaRepository<RecipeEntity, Integer> {
+public interface RecipeRepository extends JpaRepository<RecipeEntity, Long> {
     
 }
