+++ src/main/java/pe/edu/upc/MonolithFoodApplication/repositories/UserFitnessInfoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.MonolithFoodApplication.entities.UserFitnessInfoEntity;
 
 @Repository
-public interface UserFitnessInfoRepository extends JpaRepository<UserFitnessInfoEntity, Integer> {
+public interface UserFitnessInfoRepository extends JpaRepository<UserFitnessInfoEntity, Long> {
     
 }
