+++ src/main/java/pe/edu/upc/MonolithFoodApplication/repositories/UserPersonalInfoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.MonolithFoodApplication.entities.UserPersonalInfoEntity;
 
 @Repository
-public interface UserPersonalInfoRepository extends JpaRepository<UserPersonalInfoEntity, Integer> {
+public interface UserPersonalInfoRepository extends JpaRepository<UserPersonalInfoEntity, Long> {
     
 }
